@@ -1,4 +1,5 @@
-class CreateTableBooks < ActiveRecord::Migration
+migration_class = (ActiveRecord::VERSION::MAJOR >= 5) ?  ActiveRecord::Migration[4.2] : ActiveRecord::Migration
+class CreateTableBooks < migration_class
   def up
     create_table :books do |t|
       t.string :name
