@@ -1,5 +1,3 @@
-require 'apartment/deprecation'
-
 module Apartment
   module Adapters
     class AbstractAdapter
@@ -32,24 +30,6 @@ module Apartment
             yield if block_given?
           end
         end
-      end
-
-      #   Get the current tenant name
-      #
-      #   @return {String} current tenant name
-      #
-      def current_database
-        Apartment::Deprecation.warn "[Deprecation Warning] `current_database` is now deprecated, please use `current`"
-        current
-      end
-
-      #   Get the current tenant name
-      #
-      #   @return {String} current tenant name
-      #
-      def current_tenant
-        Apartment::Deprecation.warn "[Deprecation Warning] `current_tenant` is now deprecated, please use `current`"
-        current
       end
 
       #   Note alias_method here doesn't work with inheritence apparently ??
@@ -99,25 +79,14 @@ module Apartment
       #   @param {String?} tenant to connect to
       #
       def switch(tenant = nil)
-        if block_given?
-          begin
-            previous_tenant = current
-            switch!(tenant)
-            yield
-
-          ensure
-            switch!(previous_tenant) rescue reset
-          end
-        else
-          Apartment::Deprecation.warn("[Deprecation Warning] `switch` now requires a block reset to the default tenant after the block. Please use `switch!` instead if you don't want this")
+        begin
+          previous_tenant = current
           switch!(tenant)
-        end
-      end
+          yield
 
-      #   [Deprecated]
-      def process(tenant = nil, &block)
-        Apartment::Deprecation.warn("[Deprecation Warning] `process` is now deprecated. Please use `switch`")
-        switch(tenant, &block)
+        ensure
+          switch!(previous_tenant) rescue reset
+        end
       end
 
       #   Iterate over all tenants, switch to tenant and yield tenant name
@@ -147,7 +116,7 @@ module Apartment
       #
       def seed_data
         # Don't log the output of seeding the db
-        silence_warnings{ load_or_abort(Apartment.seed_data_file) } if Apartment.seed_data_file
+        silence_warnings{ load_or_raise(Apartment.seed_data_file) } if Apartment.seed_data_file
       end
       alias_method :seed, :seed_data
 
@@ -159,7 +128,7 @@ module Apartment
 
       def drop_command(conn, tenant)
         # connection.drop_database   note that drop_database will not throw an exception, so manually execute
-        conn.execute("DROP DATABASE #{environmentify(tenant)}")
+        conn.execute("DROP DATABASE #{conn.quote_table_name(environmentify(tenant))}")
       end
 
       #   Create the tenant
@@ -175,7 +144,7 @@ module Apartment
       end
 
       def create_tenant_command(conn, tenant)
-        conn.create_database(environmentify(tenant))
+        conn.create_database(environmentify(tenant), @config)
       end
 
       def connect(tenant)
@@ -195,8 +164,12 @@ module Apartment
       #   @param {String} tenant Database name
       #
       def connect_to_new(tenant)
+        query_cache_enabled = ActiveRecord::Base.connection.query_cache_enabled
+
         Apartment.connection_handler.establish_connection multi_tenantify(tenant)
         connect_to_existing(tenant)
+
+        Apartment.connection.enable_query_cache! if query_cache_enabled
       end
 
       def connect_to_existing(tenant)
@@ -228,7 +201,7 @@ module Apartment
       def import_database_schema
         ActiveRecord::Schema.verbose = false    # do not log schema load output.
 
-        load_or_abort(Apartment.database_schema_file) if Apartment.database_schema_file
+        load_or_raise(Apartment.database_schema_file) if Apartment.database_schema_file
       end
 
       #   Return a new config that is multi-tenanted
@@ -247,15 +220,17 @@ module Apartment
         config[:database] = environmentify(tenant)
       end
 
-      #   Load a file or abort if it doesn't exists
+      #   Load a file or raise error if it doesn't exists
       #
-      def load_or_abort(file)
+      def load_or_raise(file)
         if File.exists?(file)
           load(file)
         else
-          abort %{#{file} doesn't exist yet}
+          raise FileNotFound, "#{file} doesn't exist yet"
         end
       end
+      # Backward compatibility
+      alias_method :load_or_abort, :load_or_raise
 
       #   Exceptions to rescue from on db operations
       #
