@@ -18,7 +18,7 @@ module Apartment
     private
 
       def rescue_from
-        PGError
+        PG::Error
       end
     end
 
@@ -63,10 +63,17 @@ module Apartment
       #
       def connect_to_new(tenant = nil)
         return reset if tenant.nil?
-        raise ActiveRecord::StatementInvalid.new("Could not find schema #{tenant}") unless Apartment.connection.schema_exists? tenant
+        raise ActiveRecord::StatementInvalid.new("Could not find schema #{tenant}") unless Apartment.connection.schema_exists?(tenant.to_s)
 
         @current = tenant.to_s
         Apartment.connection.schema_search_path = full_search_path
+
+        # When the PostgreSQL version is < 9.3,
+        # there is a issue for prepared statement with changing search_path.
+        # https://www.postgresql.org/docs/9.3/static/sql-prepare.html
+        if postgresql_version < 90300
+          Apartment.connection.clear_cache!
+        end
 
       rescue *rescuable_exceptions
         raise TenantNotFound, "One of the following schema(s) is invalid: \"#{tenant}\" #{full_search_path}"
@@ -87,22 +94,42 @@ module Apartment
       def persistent_schemas
         [@current, Apartment.persistent_schemas].flatten
       end
+
+      def postgresql_version
+        # ActiveRecord::ConnectionAdapters::PostgreSQLAdapter#postgresql_version is
+        # public from Rails 5.0.
+        Apartment.connection.send(:postgresql_version)
+      end
     end
 
     # Another Adapter for Postgresql when using schemas and SQL
     class PostgresqlSchemaFromSqlAdapter < PostgresqlSchemaAdapter
 
       PSQL_DUMP_BLACKLISTED_STATEMENTS= [
-        /SET search_path/i,   # overridden later
-        /SET lock_timeout/i   # new in postgresql 9.3
+        /SET search_path/i,                           # overridden later
+        /SET lock_timeout/i,                          # new in postgresql 9.3
+        /SET row_security/i,                          # new in postgresql 9.5
+        /SET idle_in_transaction_session_timeout/i,   # new in postgresql 9.6
       ]
 
       def import_database_schema
-        clone_pg_schema
-        copy_schema_migrations
+        preserving_search_path do
+          clone_pg_schema
+          copy_schema_migrations
+        end
       end
 
     private
+
+      # Re-set search path after the schema is imported.
+      # Postgres now sets search path to empty before dumping the schema
+      # and it mut be reset
+      #
+      def preserving_search_path
+        search_path = Apartment.connection.execute("show search_path").first["search_path"]
+        yield
+        Apartment.connection.execute("set search_path = #{search_path}")
+      end
 
       # Clone default schema into new schema named after current tenant
       #
@@ -140,7 +167,7 @@ module Apartment
       #   @return {String} raw SQL contaning inserts with data from schema_migrations
       #
       def pg_dump_schema_migrations_data
-        with_pg_env { `pg_dump -a --inserts -t schema_migrations -t ar_internal_metadata -n #{default_tenant} #{dbname}` }
+        with_pg_env { `pg_dump -a --inserts -t #{default_tenant}.schema_migrations -t #{default_tenant}.ar_internal_metadata #{dbname}` }
       end
 
       # Temporary set Postgresql related environment variables if there are in @config
@@ -165,11 +192,21 @@ module Apartment
       def patch_search_path(sql)
         search_path = "SET search_path = \"#{current}\", #{default_tenant};"
 
-        sql
+        swap_schema_qualifier(sql)
           .split("\n")
           .select {|line| check_input_against_regexps(line, PSQL_DUMP_BLACKLISTED_STATEMENTS).empty?}
           .prepend(search_path)
           .join("\n")
+      end
+
+      def swap_schema_qualifier(sql)
+        sql.gsub(/#{default_tenant}\.\w*/) do |match|
+          if Apartment.pg_excluded_names.any? { |name| match.include? name }
+            match
+          else
+            match.gsub(default_tenant, %{"#{current}"})
+          end
+        end
       end
 
       #   Checks if any of regexps matches against input
