@@ -1,6 +1,5 @@
 require 'rack/request'
 require 'apartment/tenant'
-require 'apartment/deprecation'
 
 module Apartment
   module Elevators
@@ -10,7 +9,7 @@ module Apartment
 
       def initialize(app, processor = nil)
         @app = app
-        @processor = processor || parse_method
+        @processor = processor || method(:parse_tenant_name)
       end
 
       def call(env)
@@ -25,26 +24,8 @@ module Apartment
         end
       end
 
-      def parse_database_name(request)
-        deprecation_warning
-        parse_tenant_name(request)
-      end
-
       def parse_tenant_name(request)
         raise "Override"
-      end
-
-      def parse_method
-        if self.class.instance_methods(false).include? :parse_database_name
-          deprecation_warning
-          method(:parse_database_name)
-        else
-          method(:parse_tenant_name)
-        end
-      end
-
-      def deprecation_warning
-        Apartment::Deprecation.warn "[DEPRECATED::Apartment] Use #parse_tenant_name instead of #parse_database_name -> #{self.class.name}"
       end
     end
   end
