@@ -1,3 +1,3 @@
 module Apartment
-  VERSION = "1.2.0"
+  VERSION = "2.2.0"
 end
