@@ -17,7 +17,7 @@ module Dummy
     require 'apartment/elevators/subdomain'
     require 'apartment/elevators/domain'
 
-    config.middleware.use 'Apartment::Elevators::Subdomain'
+    config.middleware.use Apartment::Elevators::Subdomain
 
     # Custom directories with classes and modules you want to be autoloadable.
     config.autoload_paths += %W(#{config.root}/lib)
